requirements/test.txt
-r base.txt
django-jenkins
pep8

